chatbot_FastAPI/pyproject.toml
[project]
name = "chatbot-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.12",
    "fastapi[standard]>=0.115.12",
    "openai-agents[litellm,viz]>=0.0.16",
]
